requirements.txt
torch>=2.1.0
transformers>=4.34
accelerate>=0.23
peft>=0.6.1
bitsandbytes>=0.41.1
hydra-core>=1.3
deepspeed>=0.10.3
spacy>=3.7.0
scipy
rouge
jpype1
shortuuid
jsonlines
rich
wandb
gem-metrics-fork
pre-commit
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

